Adding ability to update the owner of any MainToken key -- user must provide an email address to own the key.  Still needs work. Need to make email address insert unique so multiple rows don't get added to owner.  One owner can have many maintoken keys.

+++ LibreStore/Models/MainToken.cs
@@ -1,13 +1,13 @@
 namespace LibreStore.Models;
 
 public class MainToken{
-    public int Id{get;set;}
-    public int OwnerId{get;set;}
+    public Int64 Id{get;set;}
+    public Int64 OwnerId{get;set;}
     public String Key{get; set;}
     public DateTime Created {get;set;}
     public bool Active{get;set;}
 
-    public MainToken(String key, int ownerId=0)
+    public MainToken(String key, Int64 ownerId=0)
     {
         OwnerId = ownerId;
         Key = key;
@@ -15,7 +15,7 @@ public class MainToken{
         Active = true;
     }
 
-    public MainToken(int id, String key,  DateTime created, int ownerId, bool active)
+    public MainToken(Int64 id, String key,  DateTime created, Int64 ownerId, bool active)
     {
         Id = id;
         OwnerId = ownerId;
